metrics: refactor to separate metrics package.
