+++ pkg/cke/main.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/coreos/etcd/clientv3/concurrency"
 	"github.com/cybozu-go/cke"
+	"github.com/cybozu-go/cke/metrics"
 	"github.com/cybozu-go/cke/sabakan"
 	"github.com/cybozu-go/cke/server"
 	"github.com/cybozu-go/etcdutil"
@@ -105,8 +106,8 @@ func main() {
 	well.Go(controller.Run)
 
 	// Metrics
-	collector := cke.NewCollector(etcd)
-	metricsHandler := cke.GetHandler(collector)
+	collector := metrics.NewCollector(etcd)
+	metricsHandler := metrics.GetHandler(collector)
 	mux := http.NewServeMux()
 	mux.Handle("/metrics", metricsHandler)
 	ms := &well.HTTPServer{
